+++ apps/dapp/src/components/Pages/Ascend/components/Trade/hooks/use-vault-trade-state.ts
@@ -1,15 +1,15 @@
-import { useEffect, useReducer, useState } from 'react';
-import { BigNumber, Contract } from 'ethers';
-import { useBalance, useContractReads } from 'wagmi';
+import { useReducer } from 'react';
+import { BigNumber } from 'ethers';
+import { useBalance } from 'wagmi';
 
-import balancerPoolAbi from 'data/abis/balancerPool.json';
-import balancerVaultAbi from 'data/abis/balancerVault.json';
 import { Pool } from 'components/Layouts/Ascend/types';
 import { useWallet } from 'providers/WalletProvider';
 import { ZERO } from 'utils/bigNumber';
 import { getBigNumberFromString } from 'components/Vault/utils';
 import { Nullable } from 'types/util';
-import { getSwapLimit, getSwapDeadline } from './utils';
+import { getSwapLimit, getSwapDeadline } from '../utils';
+
+import { useVaultContract } from './use-vault-contract';
 
 type Action<A extends ActionType, P extends any> = { type: A, payload: P };
 
@@ -19,7 +19,7 @@ enum ActionType {
   SetSwapQuote,
   SetSwapQuoteLoading,
   SetTransactionSettings,
-  ResetSwapState,
+  ResetQuoteState,
   UpdateSwapState,
 };
 
@@ -29,7 +29,7 @@ type Actions =
   Action<ActionType.SetSwapQuote, Nullable<BigNumber>> |
   Action<ActionType.SetSwapQuoteLoading, boolean> |
   Action<ActionType.SetTransactionSettings, TradeState['transactionSettings']> |
-  Action<ActionType.ResetSwapState, null> |
+  Action<ActionType.ResetQuoteState, null> |
   Action<ActionType.UpdateSwapState, { isLoading: boolean; error: string }>;
 
 interface TradeState {
@@ -73,7 +73,7 @@ const reducer = (state: TradeState, action: Actions): TradeState => {
         },
       };
     }
-    case ActionType.ResetSwapState: {
+    case ActionType.ResetQuoteState: {
       return {
         ...state,
         sell: {
@@ -141,82 +141,6 @@ const reducer = (state: TradeState, action: Actions): TradeState => {
   }
 };
 
-const useVaultContract = (pool: Pool) => {
-  const { wallet, signer } = useWallet();
-  const [vaultContract, setVaultContract] = useState<Contract>();
-
-  const { data } = useContractReads({
-    contracts: [{
-      addressOrName: pool.address,
-      contractInterface: balancerPoolAbi,
-      functionName: 'getVault',
-    }],
-  });
-
-  useEffect(() => {
-    if (vaultContract || !data || !signer) {
-      return;
-    }
-
-    const vaultAddress = !!data && data.length > 0 ? data[0] : '';
-    setVaultContract(new Contract(vaultAddress as string, balancerVaultAbi, signer));
-  }, [data, vaultContract, signer, setVaultContract]);
-
-  return {
-    address: vaultContract?.address || '',
-    isReady: !!vaultContract && !!wallet,
-    async getSwapQuote(amount: BigNumber, sellAssetAddress: string, buyAssetAddress: string) {
-      const assetOutIndex = pool.tokensList.findIndex((address) => address === buyAssetAddress);
-      const assetInIndex = pool.tokensList.findIndex((address) => address === sellAssetAddress);
-
-      return vaultContract!.callStatic.queryBatchSwap(
-        0,
-        [{
-          poolId: pool.id,
-          assetInIndex,
-          assetOutIndex,
-          amount,
-          userData: '0x',
-        }],
-        pool.tokensList,
-        {
-          sender: wallet!.toLowerCase(),
-          recipient: wallet!.toLowerCase(),
-          fromInternalBalance: false,
-          toInternalBalance: false,
-        },
-      );
-    },
-    async swap(
-      amount: BigNumber,
-      sellAssetAddress: string,
-      buyAssetAddress: string,
-      limits: BigNumber,
-      deadline: BigNumber,
-    ) {
-      const swap = {
-        poolId: pool.id,
-        kind: 0,
-        assetIn: sellAssetAddress,
-        assetOut: buyAssetAddress,
-        amount,
-        userData: '0x',
-      };
-      
-      const funds = {
-        sender: wallet!.toLowerCase(),
-        recipient: wallet!.toLowerCase(),
-        fromInternalBalance: false,
-        toInternalBalance: false,
-      };
-
-      return vaultContract!.swap(swap, funds, limits, deadline, {
-        gasLimit: 400000,
-      });
-    },
-  };
-};
-
 export const useVaultTradeState = (pool: Pool) => {
   const { wallet } = useWallet();
   const vaultContract = useVaultContract(pool);
@@ -239,7 +163,7 @@ export const useVaultTradeState = (pool: Pool) => {
     quote: {
       loading: false,
       estimate: null,
-      estimateWithSlippage: ZERO,
+      estimateWithSlippage: null,
     },
     swap: {
       error: '',
@@ -247,7 +171,7 @@ export const useVaultTradeState = (pool: Pool) => {
     },
     transactionSettings: {
       slippageTolerance: 1,
-      deadlineMinutes: 20,
+      deadlineMinutes: 5,
     },
   });
 
@@ -323,7 +247,10 @@ export const useVaultTradeState = (pool: Pool) => {
         state.quote.estimateWithSlippage!,
         deadline,
       );
+      
       dispatch({ type: ActionType.UpdateSwapState, payload: { isLoading: false, error: '' } });
+      dispatch({ type: ActionType.ResetQuoteState, payload: null });
+
     } catch (err) {
       console.error('Error swapping', err)
       dispatch({ type: ActionType.UpdateSwapState, payload: { isLoading: false, error: (err as Error).message } });
